Update devcontainer config to build plugin
Changing devcontainer.config to build an ubuntu-focal
image from scratch and install java 17.0.9-amzn using SDKMan.
Image also creates a non-root user called 'infino' to as opensearch
cannot be run as root user. Testing the infino-opensearch-plugin
spins up a snapshot of opensearch and this fails if the test is
run as root.