OpenSearch Plugin: Change HttpClient::sendAsync to synchronous HttpClient::send to get around the java security policy exception (#244)
The PR changes the async http call made to Infino server to a sync call. The async call was spawning off new threads which were being blocked by opensearch's strict security policy that forbids thread operations. Changing the call to a synchronous send does not trigger the security policy exception.

Changing the async call to sync is ok since each sync call is still made in a separate thread. So the plugin itself will not block on send() call.

# Curious case of threads in OpenSearch

The thread that we manually spawn off in the plugin in InfinoRestHandler (when processing incoming requests) seems to work fine. However the thread that HttpClient::sendAsync() spawns internally runs into OpenSearch's security manager rules. My hypotheses here is that OpenSearch's security manager grants access to modify only the current thread group or the current thread group's parent - check this comment in OpenSearch source code and the sendAsync() call is perhaps running in a different thread group context and is trying to access/modify OpenSearch's threadgroup. Still not fully clear and will need further investigation. But for now since we have a tested workaround, we'll go ahead with this to make progress.