docker-compose config to deploy Infino, OpenSearch + Infino plugin and OpenSearch Dasbhoard (#248)
The PR adds a docker-compose.yml file that deploys and configures the below containers
1. OpenSearch dashboard
2. OpenSearch with Infino plugin
3. Infino server

Readme file has been updated with the necessary documentation for using the Infino via the OpenSearch dashboard after deploying the docker-compose file. I have also updated the Readme.md file to move around some sections and add a table of contents (as the Readme is now beginning to grow longer).